Add automatic daily database exports to local fs
This commit adds the https://github.com/jan-di/docker-database-backup
container. This is a tool to exort the contents of running database
containers to *.sql files. This allows for external backups of the state
of the cluster for disaster recovery purposes (or for setting up test
servers).

Backups are written to /dump on the backup service containers, which is
mounted as ./state/db-backups via Docker Compose.